refactor : 토큰 생성 및 유효성 검증 함수 수정 1. 파일 위치를 /utils에서 /utils/token로 이동 2. catch문에서 에러 처리 함수를 사용하여 에러를 처리하도록 수정
